requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# Deep learning
torch==2.0.1
dgl==1.1.1

# Data processing
pyarrow==14.0.1
fastparquet==2023.7.0

# Web API
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
werkzeug==2.3.7

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0

# Kafka
kafka-python==2.0.2
confluent-kafka==2.2.0

# Configuration and logging
pyyaml==6.0.1
python-dotenv==1.0.0

# Metrics and monitoring
prometheus-client==0.16.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development tools
black==23.7.0
isort==5.12.0
flake8==6.1.0

# Production tools
uvicorn==0.23.2
httpx==0.24.1
tenacity==8.2.3
python-json-logger==2.0.7

# Additional dependencies for API and frontend
fastapi==0.108.0
pydantic==2.5.3
requests==2.31.0
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
pytest-cov==4.1.0
pyyaml==6.0.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
lightfm==1.17 